+++ R/prompt.R
@@ -111,7 +111,16 @@ unitizer_prompt <- function(
   )
   repeat {
     prompt.txt <- sprintf("%s> ", "unitizer")
-    val <- tryCatch(faux_prompt(prompt.txt), simpleError=function(e) e)
+
+    old.opt <- options(warn=1)
+    on.exit(options(old.opt))
+    val <- tryCatch(
+      faux_prompt(prompt.txt),
+      simpleError=function(e) e
+    )
+    on.exit(NULL)
+    options(old.opt)
+
     if(inherits(val, "simpleError")) {
       cond.chr <- as.character(val)
       cat(cond.chr, file=stderr())
@@ -366,7 +375,7 @@ exit_fun <- function(y, env, valid.vals) {
 read_line <- function(prompt="") {
   stopifnot(is.chr1(prompt))
   if(is.null(.global$prompt.vals)) {
-    readline(prompt)
+    readline(prompt)  # nocov can't test this in non-interactive
   } else if(!is.character(.global$prompt.vals)) {
     stop( # nocov start
       "Internal Error: internal object `.global$prompt.vals` has unexpected ",