+++ tests/testthat/helper/runtt.R
@@ -57,6 +57,9 @@ local({
   old.opt.2 <- c(
     old.opt.2, options(unitizer.state='recommended', diffobj.pager='off')
   )
+  # A dummy class we use to test coercion errors
+
+  setClass("uhtsdfoqiuerhzb", slots=c(a='integer'))
 
   on.exit(
     {
@@ -67,43 +70,49 @@ local({
   )
   # Run tests
 
-  test.res <- test_dir(
-    "testthat",
-    reporter="check",
-    env=environment(),
-    filter=paste(sep="|",
-      "browse",
-      "capture",
-      "change",
-      "demo",
-      "error",
-      "exec",
-      "get",
-      "global",
-      "handledruns",
-      "inpkg",
-      "ischecks",
-      "item",
-      "list",
-      "misc",
-      "options",
-      "parse",
-      "prompt",
-      "rename",
-      "repairenvs",
-      "search",
-      "section",
-      "shim",
-      "state",
-      "text",
-      "translate",
-      "utz1",
-      "utz2",
-      "upgrade",
-      "zzrunlast"
-    )
+  test.filt <- paste(sep="|",
+    "browse",
+    "capture",
+    "change",
+    "demo",
+    "error",
+    "exec",
+    "get",
+    "global",
+    "handledruns",
+    "inpkg",
+    "ischecks",
+    "item",
+    "list",
+    "misc",
+    "options",
+    "parse",
+    "prompt",
+    "rename",
+    "repairenvs",
+    "random",
+    "search",
+    "section",
+    "shim",
+    "state",
+    "text",
+    "translate",
+    "utz1",
+    "utz2",
+    "upgrade",
+    "zzrunlast"
   )
-
+  test.res <- if(packageVersion('testthat') > "1.0.2") {
+    test_dir(
+      "testthat",
+      reporter="check",
+      env=environment(),
+      filter=test.filt,
+      wrap=FALSE
+    )
+  } else {
+    test_dir("testthat", reporter="check", env=environment(), filter=test.filt)
+  }
   # Check for failures and throw error if they exist since test_dir does not
   # do so on its own